Refactor stackable and non-stackable Flashbar into different files
