+++ src/flashbar/non-stackable-flashbar.tsx
@@ -1,0 +1,153 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+import clsx from 'clsx';
+import React, { useEffect, useMemo, useRef, useState } from 'react';
+import { Flash, focusFlashById } from './flash';
+import { FlashbarProps } from './interfaces';
+import { getBaseProps } from '../internal/base-component';
+import { TIMEOUT_FOR_ENTERING_ANIMATION } from './constant';
+import { TransitionGroup } from 'react-transition-group';
+import { Transition } from '../internal/components/transition';
+import useBaseComponent from '../internal/hooks/use-base-component';
+import { useContainerBreakpoints } from '../internal/hooks/container-queries';
+import { useMergeRefs } from '../internal/hooks/use-merge-refs';
+import { useReducedMotion, useVisualRefresh } from '../internal/hooks/use-visual-mode';
+import { getVisualContextClassname } from '../internal/components/visual-context';
+
+import styles from './styles.css.js';
+
+export { FlashbarProps };
+
+export default function NonStackableFlashbar({ items, ...restProps }: FlashbarProps) {
+  const { __internalRootRef } = useBaseComponent('Flashbar');
+  const baseProps = getBaseProps(restProps);
+
+  const ref = useRef<HTMLDivElement | null>(null);
+  const [breakpoint, breakpointRef] = useContainerBreakpoints(['xs']);
+  const mergedRef = useMergeRefs(ref, breakpointRef, __internalRootRef);
+
+  const isVisualRefresh = useVisualRefresh();
+
+  const [previousItems, setPreviousItems] = useState<ReadonlyArray<FlashbarProps.MessageDefinition>>(items);
+  const [nextFocusId, setNextFocusId] = useState<string | null>(null);
+
+  const isReducedMotion = useReducedMotion(breakpointRef as any);
+  const allItemsHaveId = useMemo(() => items.every(item => 'id' in item), [items]);
+
+  // Track new or removed item IDs in state to only trigger focus changes for newly added items.
+  // https://reactjs.org/docs/hooks-faq.html#how-do-i-implement-getderivedstatefromprops
+  if (items) {
+    const newItems = items.filter(({ id }) => id && !previousItems.some(item => item.id === id));
+    const removedItems = previousItems.filter(({ id }) => id && !items.some(item => item.id === id));
+    if (newItems.length > 0 || removedItems.length > 0) {
+      setPreviousItems(items);
+      const newFocusItems = newItems.filter(({ ariaRole }) => ariaRole === 'alert');
+      if (newFocusItems.length > 0) {
+        setNextFocusId(newFocusItems[0].id!);
+      }
+    }
+  }
+
+  useEffect(() => {
+    if (nextFocusId) {
+      focusFlashById(ref.current, nextFocusId);
+    }
+  }, [nextFocusId]);
+
+  /**
+   * All the flash items should have ids so we can identify which DOM element is being
+   * removed from the DOM to animate it. Motion will be disabled if any of the provided
+   * flash messages does not contain an `id`.
+   */
+  const motionDisabled = isReducedMotion || !isVisualRefresh || !allItemsHaveId;
+
+  const animateFlash = !isReducedMotion && isVisualRefresh;
+
+  /**
+   * If the flashbar is flat and motion is `enabled` then the adding and removing of items
+   * from the flashbar will render with visual transitions.
+   */
+  function renderFlatItemsWithTransitions() {
+    if (motionDisabled || !items) {
+      return;
+    }
+
+    return (
+      // This is a proxy for <ul>, so we're not applying a class to another actual component.
+      // eslint-disable-next-line react/forbid-component-props
+      <TransitionGroup component="ul" className={styles['flash-list']}>
+        {items &&
+          items.map((item, index) => (
+            <Transition
+              transitionChangeDelay={{ entering: TIMEOUT_FOR_ENTERING_ANIMATION }}
+              key={item.id ?? index}
+              in={true}
+            >
+              {(state: string, transitionRootElement: React.Ref<HTMLDivElement> | undefined) => (
+                <li className={styles['flash-list-item']}>
+                  {renderItem(item, item.id ?? index, transitionRootElement, state)}
+                </li>
+              )}
+            </Transition>
+          ))}
+      </TransitionGroup>
+    );
+  }
+
+  /**
+   * If the flashbar is flat and motion is `disabled` then the adding and removing of items
+   * from the flashbar will render without visual transitions.
+   */
+  function renderFlatItemsWithoutTransitions() {
+    if (!motionDisabled || !items) {
+      return;
+    }
+
+    return (
+      <ul className={styles['flash-list']}>
+        {items.map((item, index) => (
+          <li key={item.id ?? index} className={styles['flash-list-item']}>
+            {renderItem(item, item.id ?? index)}
+          </li>
+        ))}
+      </ul>
+    );
+  }
+
+  /**
+   * This is a shared render function for a single flashbar item to be used
+   * by the stacking, motion, and non-motion item group render functions.
+   */
+  function renderItem(
+    item: FlashbarProps.MessageDefinition,
+    key: string | number,
+    transitionRootElement?: React.Ref<HTMLDivElement> | undefined,
+    transitionState?: string | undefined
+  ) {
+    return (
+      <Flash
+        // eslint-disable-next-line react/forbid-component-props
+        className={clsx(
+          getVisualContextClassname('flashbar'),
+          animateFlash && styles['flash-with-motion'],
+          isVisualRefresh && styles['flash-refresh']
+        )}
+        key={key}
+        ref={transitionRootElement}
+        transitionState={transitionState}
+        {...item}
+      />
+    );
+  }
+
+  return (
+    <div
+      {...baseProps}
+      className={clsx(baseProps.className, styles.flashbar, styles[`breakpoint-${breakpoint}`])}
+      ref={mergedRef}
+    >
+      {renderFlatItemsWithTransitions()}
+      {renderFlatItemsWithoutTransitions()}
+    </div>
+  );
+}
